Add scripts, Move catcher to module
