+++ package.json
@@ -6,9 +6,11 @@
   "main": "dist/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "prestart": "npm run-script build",
+    "prebuild": "rm -rf dist",
     "build": "tsc",
-    "start": "node --enable-source-maps ."
+    "prestart": "npm run-script build",
+    "start": "node --enable-source-maps .",
+    "watch": "nodemon -e ts,tsx --watch src --exec \"npm run start\""
   },
   "author": "TheMas3212",
   "license": "MIT",
@@ -18,6 +20,7 @@
     "@typescript-eslint/eslint-plugin": "^8.14.0",
     "@typescript-eslint/parser": "^8.14.0",
     "eslint": "^9.14.0",
+    "nodemon": "^3.1.7",
     "typescript": "^5.6.3"
   }
 }
