+++ src/modules/catcher.ts
@@ -87,3 +87,10 @@ export function catcher(shellPort: number, shellHost: string, ipcPort: number) {
   console.log(`Listening for Reverse Shells on ${shellHost}:${shellPort}`);
   console.log(`IPC on ${ipcPort}`);
 }
+
+// Check if the file was imported or run directly.
+if (require.main === module) {
+  // Directly run from CLI.  Collect parameters and start catcher.
+  const [shellPort, shellHost, ipcPort]= process.argv.slice(2,5)
+  catcher(Number(shellPort), shellHost, Number(ipcPort));
+}
